Reduced size of integers passed to serial comm functions due to performance reasons

+++ src/libs/serial_communication/serial_communication.h
@@ -5,8 +5,8 @@
 #define NEWLINE "\n\r"
 
 void serialInit(uint32_t ubrr);
-void writeUint(uint64_t n);
-void writeBinary(uint64_t n);
+void writeUint(uint16_t n);
+void writeBinary(uint16_t n);
 void writeByte(const char c);
 void writeString(const char *str);
 void writeStringF(const char *format, ...);
